pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "video-qa"
version = "0.1.0"
description = "Analyze screen recordings to auto-detect key moments, generate timeline comparisons, and create test reports"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "miyashita337" }]
keywords = ["video", "qa", "testing", "screen-recording", "visual-regression"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9.0",
    "pillow>=10.0.0",
]

[project.scripts]
video-qa = "video_qa.cli:app"

[project.urls]
Homepage = "https://github.com/miyashita337/video-qa"
Repository = "https://github.com/miyashita337/video-qa"
Issues = "https://github.com/miyashita337/video-qa/issues"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/video_qa"]
